requirements.txt
aniso8601==9.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
click==8.0.1
colorama==0.4.4
dnspython==1.16.0
Flask==2.0.1
flask-restx==0.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
pymongo==3.12.0
pyrsistent==0.18.0
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
Werkzeug==2.0.1
